Removed extraneous/outdated scripts, moved analysis tool into analysis/ subdir in /, and added a message to inform users about whether the analysis tool will be built at "cmake time"
