+++ analysis/parallelPdfAnalysis.py
@@ -8,6 +8,7 @@ from decimal import *
 import math
 import re
 import networkx as nx
+#import pydot
 import os
 import struct
 import matplotlib
@@ -16,7 +17,6 @@ import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 import plfsinterface
 import analysis
-import multiprocessing
 import graphingLibrary
 import gc
 import operator
@@ -86,58 +86,40 @@ def graphPerProcessor(mpiFile,numberOfProcessors, average, jobID, above, below):
                inputList.append((id, beg, end))
     gc.enable()
     timeFile.close()
-    inputList.sort( key = operator.itemgetter(0,1))
-    while len(inputList) > (100*numberOfProcessors) or len(inputList) > 65536:
-        inputList = inputList[::2] # get every other element
-    low = min(x[1] for x in inputList)
-    high = max(x[2] for x in inputList)
-    ax2.set_ylim([low,high])
-    ax2.plot([[x[0] for x in inputList], [x[0] for x in inputList]],
-        [[x[1] for x in inputList], [x[2] for x in inputList]], linewidth=1.5)
+    # sometimes we have nothing to plot if no procs are outside one std dev.
+    if (len(inputList) > 0):
+        inputList.sort( key = operator.itemgetter(0,1))
+        while len(inputList) > (100*numberOfProcessors) or len(inputList) > 65536:
+            inputList = inputList[::2] # get every other element
+        low = min(x[1] for x in inputList)
+        high = max(x[2] for x in inputList)
+        ax2.set_ylim([low,high])
+        ax2.plot([[x[0] for x in inputList], [x[0] for x in inputList]],
+            [[x[1] for x in inputList], [x[2] for x in inputList]], linewidth=1.5)
 
-
-def graph((function, times, bandwidths, iosTime, iosFin, writeBins, \
-           hostdirs, sizes, mpiFile, average, jobID, \
-           above, below, name)):
-    units = {0: " B", 1:" KiB", 2:" MiB", 3:" GiB", 4:" TiB", 5:" PiB"}
-    if function == graphingLibrary.bandwidthGraphs:
-        graphingLibrary.bandwidthGraphs(times, bandwidths, iosTime, iosFin,\
-            hostdirs, sizes, units)
-    elif function == graphingLibrary.networkGraph:
-        graphingLibrary.networkGraph(hostdirs)
-    elif function == graphingLibrary.barGraph:
-        graphingLibrary.barGraph(writeBins, units)
-    elif function == graphPerProcessor:
-        graphPerProcessor(mpiFile, len(hostdirs)-1, average, jobID, above, \
-            below)
-    logicalFile = hostdirs[0][1]
-    logicalFile = logicalFile.replace("/", "_")
-    plt.savefig('Analysis%s%s.pdf' % (logicalFile, name))
-    plt.close()
-
-# gives each processor a graph to output 
 def generateGraphs(times, bandwidths, iosTime, iosFin, writeBins,\
                     hostdirs, sizes, processorGraphs, mpiFile, average, \
                     jobID, above, below):
     matplotlib.rc("xtick", labelsize=10)
     matplotlib.rc("ytick", labelsize=10)
-    pool = multiprocessing.Pool(4)
+    logicalFilename = hostdirs[0][1]
+    logicalFilename = logicalFilename.replace("/", "_")
+    pdf = PdfPages("Analysis" + logicalFilename + ".pdf")
+    #cooresponds to the scale functions count, if more units added,
+    #scale should be adjusted as well
+    units = {0: " B", 1:" KiB", 2:" MiB", 3:" GiB", 4:" TiB", 5:" PiB"}
+    graphingLibrary.bandwidthGraphs(times, bandwidths, iosTime, iosFin, \
+        hostdirs, sizes, units)
+    pdf.savefig()
+    plt.close()
+    graphingLibrary.networkGraph(hostdirs)
+    pdf.savefig()
+    plt.close()
+    graphingLibrary.barGraph(writeBins, units)
+    pdf.savefig()
+    plt.close()
     if processorGraphs:
-        input = [(graphingLibrary.bandwidthGraphs, times, bandwidths, iosTime,\
-            iosFin, None, hostdirs, sizes, None, None, None, None, None, \
-            "-Bandwidths"), (graphingLibrary.networkGraph, None, None, None, \
-            None, None, hostdirs, None, None, None, None, None, None,\
-            "-Network"), (graphingLibrary.barGraph, None, None, None, None, \
-            writeBins, hostdirs, None, None, None, None, None, None, \
-            "-WriteSizes"), (graphPerProcessor, None, None, None, None, None, \
-            hostdirs, None, mpiFile, average, jobID, above, below, \
-            "-Processors")]
-    else:
-        input = [(graphingLibrary.bandwidthGraphs, times, bandwidths, iosTime, \
-            iosFin, None, hostdirs, sizes, None, None, None, None, None, \
-            "-Bandwidths"), (graphingLibrary.networkGraph, None, None, None, \
-            None, None, hostdirs, None, None, None, None, None, None,\
-            "-Network"), (graphingLibrary.barGraph, None, None, None, None, \
-            writeBins, hostdirs, None, None, None, None, None, None, \
-            "-WriteSizes")]
-    pool.map(graph, input)
+        graphPerProcessor(mpiFile,len(hostdirs)-1, average, jobID, above, below)
+        pdf.savefig()
+        plt.close()
+    pdf.close()
